requirements-test.txt
alabaster==0.7.16
alembic==1.16.2
aniso8601==10.0.1
attrs==25.3.0
autoflake==1.7.7
aws-wsgi==0.2.7
babel==2.17.0
bandit==1.8.5
black==24.10.0
blinker==1.9.0
boto3==1.38.46
botocore==1.38.46
build==1.2.2.post1
cachelib==0.13.0
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.9.1
cssbeautifier==1.15.4
Deprecated==1.2.18
distlib==0.3.9
djlint==1.36.4
docutils==0.20.1
dominate==2.9.1
dparse==0.6.4
EditorConfig==0.17.1
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==25.9.2
filelock==3.18.0
flake8==7.3.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.1
Flask-Cors==4.0.2
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
identify==2.6.12
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.0
limits==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
-e git+https://github.com/nivecher/meal-expense-tracker.git@6c2a8912ee6e5c2f0649b6d4d2f70a310a390ad4#egg=meal_expense_tracker
msgspec==0.19.0
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
openpyxl==3.1.5
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.4.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==13.9.4
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
safety==2.3.4
selenium==4.34.0
setuptools==80.9.0
setuptools-scm==8.3.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stevedore==5.4.1
toml==0.10.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
urllib3==2.4.0
virtualenv==20.31.2
visitor==0.1.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
